--- a/core-ducktape/src/main/java/net/year4000/utilities/ducktape/module/internal/ModuleHandle.java
+++ b/core-ducktape/src/main/java/net/year4000/utilities/ducktape/module/internal/ModuleHandle.java
@@ -6,6 +6,9 @@ package net.year4000.utilities.ducktape.module.internal;
 /** This is the base interface that the module will inherit */
 public interface ModuleHandle {
 
+    /** This will return the class of the loaded module */
+    Class<?> $class();
+
     /** When this is invoked it will return the instance that the proxy has */
     Object $this();
 }