--- a/core-ducktape/src/main/java/net/year4000/utilities/ducktape/module/internal/ModuleInvocationHandler.java
+++ b/core-ducktape/src/main/java/net/year4000/utilities/ducktape/module/internal/ModuleInvocationHandler.java
@@ -9,7 +9,6 @@ import net.year4000.utilities.ErrorReporter;
 import net.year4000.utilities.ducktape.module.*;
 import net.year4000.utilities.reflection.MethodHandler;
 import net.year4000.utilities.reflection.Reflections;
-import net.year4000.utilities.tuple.Pair;
 
 import java.lang.annotation.Annotation;
 import java.lang.invoke.MethodHandle;
@@ -24,12 +23,13 @@ import java.util.Map;
 /** This is the custom invocation handler that handles the module without directly invoking its methods */
 public class ModuleInvocationHandler implements InvocationHandler {
     /** The custom Lookup instance that allows invoking of private methods and fields, see BlackMagicTest for more */
-    private static final MethodHandles.Lookup LOOKUP = Reflections.<MethodHandles.Lookup>getter(MethodHandles.Lookup.class, MethodHandles.publicLookup(), "IMPL_LOOKUP").getOrThrow();
-    private static final ImmutableMap<Class<? extends Annotation>, Pair<String, Class<?>>> METHOD_MAP = ImmutableMap.of(
-        Load.class, new Pair<>("load", Loader.class),
-        Enable.class, new Pair<>("enable", Enabler.class)
+    private static final MethodHandles.Lookup lookup = Reflections.<MethodHandles.Lookup>getter(MethodHandles.Lookup.class, MethodHandles.publicLookup(), "IMPL_LOOKUP").getOrThrow();
+    // Key is the annotation that maps to the value of a functional interface class
+    private static final ImmutableMap<Class<? extends Annotation>, Class<?>> methodMap = ImmutableMap.of(
+        Load.class, Loader.class,
+        Enable.class, Enabler.class
     );
-    private final Map<String, MethodHandler> methodLookup = new HashMap<>();
+    private final Map<Method, MethodHandler> methodLookup = new HashMap<>();
     private final Object moduleInstance;
 
     private ModuleInvocationHandler(Object moduleInstance) {
@@ -40,29 +40,30 @@ public class ModuleInvocationHandler implements InvocationHandler {
     public static Object createProxy(Class<?> moduleClass, Object moduleInstance) {
         final ModuleInvocationHandler invocationHandler = new ModuleInvocationHandler(moduleInstance);
         final List<Class<?>> interfaces = new ArrayList<>();
-        // Add our custom method into the proxy
-        invocationHandler.methodLookup.put("$this", args -> moduleInstance);
         // Add the methods that we want into the proxy class with the cached method handler
         for (Method method : moduleClass.getMethods()) {
-            METHOD_MAP.forEach((annotation, pair) -> {
+            ModuleInvocationHandler.methodMap.forEach((annotation, methodInterface) -> {
                 if (method.isAnnotationPresent(annotation)) {
                     try {
-                        MethodHandle handle = ModuleInvocationHandler.LOOKUP.unreflect(method).bindTo(moduleInstance);
-                        interfaces.add(pair.b.get());
-                        invocationHandler.methodLookup.put(pair.a.get(), handle::invokeWithArguments);
+                        MethodHandle handle = ModuleInvocationHandler.lookup.unreflect(method).bindTo(moduleInstance);
+                        interfaces.add(methodInterface);
+                        // Functional interfaces should only have one method
+                        invocationHandler.methodLookup.put(methodInterface.getMethods()[0], handle::invokeWithArguments);
                     } catch (IllegalAccessException error) {
                         throw ErrorReporter.builder(error).buildAndReport();
                     }
                 }
             });
         }
+        // Add our custom methods into the proxy instance ModuleHandle.class
+        invocationHandler.methodLookup.put(Reflections.method(ModuleHandle.class, "$this").get(), args -> moduleInstance);
         return Reflections.proxy(ModuleHandle.class, invocationHandler, interfaces);
     }
 
     @Override
     public Object invoke(Object proxy, Method method, Object[] args) {
         try {
-            MethodHandler proxyMethod = this.methodLookup.get(method.getName());
+            MethodHandler proxyMethod = this.methodLookup.get(method);
             if (proxyMethod != null) {
                 return proxyMethod.handle(args);
             }