--- a/core-ducktape/src/main/java/net/year4000/utilities/ducktape/module/internal/ModuleInvocationHandler.java
+++ b/core-ducktape/src/main/java/net/year4000/utilities/ducktape/module/internal/ModuleInvocationHandler.java
@@ -57,6 +57,7 @@ public class ModuleInvocationHandler implements InvocationHandler {
         }
         // Add our custom methods into the proxy instance ModuleHandle.class
         invocationHandler.methodLookup.put(Reflections.method(ModuleHandle.class, "$this").get(), args -> moduleInstance);
+        invocationHandler.methodLookup.put(Reflections.method(ModuleHandle.class, "$class").get(), args -> moduleClass);
         return Reflections.proxy(ModuleHandle.class, invocationHandler, interfaces);
     }