--- a/core/src/main/java/net/year4000/utilities/annotations/Nullable.java
+++ b/core/src/main/java/net/year4000/utilities/annotations/Nullable.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright 2019 Year4000. All Rights Reserved.
+ */
+package net.year4000.utilities.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/** This notes that the field can be null at some point */
+@Retention(RetentionPolicy.SOURCE)
+@Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER, ElementType.LOCAL_VARIABLE})
+public @interface Nullable {
+}