--- a/core/src/main/java/net/year4000/utilities/reflection/MethodHandler.java
+++ b/core/src/main/java/net/year4000/utilities/reflection/MethodHandler.java
@@ -1,12 +1,11 @@
 /*
- * Copyright 2016 Year4000. All Rights Reserved.
+ * Copyright 2019 Year4000. All Rights Reserved.
  */
-
 package net.year4000.utilities.reflection;
 
 /** A interface to cache the method and avoid reflection look up calls */
 @FunctionalInterface
 public interface MethodHandler {
     /** Handle the method args */
-    Object handle(Object instance, Object[] args) throws Throwable;
+    Object handle(Object... args) throws Throwable;
 }