… core/src/main/java/net/year4000/utilities/reflection/MethodHandleHandlers.java
… core/src/main/java/net/year4000/utilities/reflection/ReflectionHandlers.java
… core/src/main/java/net/year4000/utilities/reflection/annotations/Getter.java
… core/src/main/java/net/year4000/utilities/reflection/annotations/Invoke.java
… core/src/main/java/net/year4000/utilities/reflection/annotations/Proxied.java
… core/src/main/java/net/year4000/utilities/reflection/annotations/Setter.java
… core/src/main/java/net/year4000/utilities/reflection/annotations/Static.java
… core/src/main/java/net/year4000/utilities/reflection/lookups/AbstractSignatureLookup.java
… core/src/main/java/net/year4000/utilities/reflection/lookups/FieldSignatureLookup.java
… core/src/main/java/net/year4000/utilities/reflection/lookups/MethodSignatureLookup.java
… core/src/test/java/net/year4000/utilities/reflection/BlackMagicTest.java
… core/src/test/java/net/year4000/utilities/reflection/InvokeTest.java
… core/src/test/java/net/year4000/utilities/reflection/ReflectionTest.java