This will use the MethodHandle api for the Proxy system instead of reflection