@ewized 2019-07-17
Signature based lookup still uses reflection based lookup, I'm fine with this right now.