@ewized 2019-07-17
Test Results for Performance Increase, was ran using the timeTest unit test method.

Ran on: 
```
MacBook Pro (Retina, 13-inch, Early 2015)
3.1 GHz Intel Core i7
16 GB 1867 MHz DDR3
```

Before Pull Request:
- 3s 948ms
- 3s 361ms
- 3s 448ms
- 3s 716ms
- 4s 248ms

After Pull Request:
- 7s 395ms
- 5s 812ms
- 5s 473ms
- 5s 783ms
- 5s 42ms