@ewized 2019-07-17
Cause of the reduction of performance is using `invoke` and not `invokeExact`. This needs to be fixed and retested before merging.