@ewized 2019-07-20
Would merge if there is an option to switch between methodhandles and reflection. Not sure if it should be the proxy level in `@Proxied(methodHandle=true)` or in the individual methods via `@Invoke(methodHandle=true)`